--- webapp/src/application/Entity/index.ts
@@ -1,1 +1,0 @@
-export { default as Factory } from './Factory';
