+++ webapp/src/domain/Entity/index.ts
@@ -1,1 +1,2 @@
+export { default as entityFactory } from './Factory';
 export { default as Entity } from './Entity';
