+++ webapp/src/infrastructure/APIClient/index.ts
@@ -1,0 +1,1 @@
+export  { getEntity, getEntityList, getKindList, getProjectList } from './APIClient'
