+++ webapp/src/presentation/views/EntityList/EntityList.tsx
@@ -5,6 +5,7 @@ import { EntityListBody } from './components/EntityListBody';
 
 
 export default function EntityList() {
+
     return (
         <div className={'EntityList'}>
             <MenuBar/>
