+++ webapp/src/presentation/views/EntityList/components/EntityListBody/EntityListBody.tsx
@@ -290,6 +290,8 @@ const useStyles = makeStyles((theme: Theme) =>
   }),
 );
 
+
+
 export default function EnhancedTable() {
   const classes = useStyles();
   const [order, setOrder] = React.useState<Order>('asc');
